// src/types/SkeltonType.ts
export type SkeletonType =
  | "slider"
  | "product"
  | "category"
  | "blog"
  | "feature"
  | "hero"
  | "service"
  | "image"
  | "contact"
  | "payment"
  | "faq"
  | "profile"
  | "cart-table"
  | "table"
  | "custome"
  | "about"
  | "brand"
  | "list"
  | "blogDetails"
  | "order-card"
  | "productDetails"
  | "coupon"
  | "footer"
  | "testimonail"
  | "navbar"
  | "procedure-section"
  | "procedure-card"
  | "stats"
  | "procedure-section-home"
  | "faq-section"
  | "about-home"
  | "modern-tech"
  | "pre-visit"
  | "patient-education-home-section"
  | "contact-section"
  | "success-stories"
  | "procedure-intro"
  | "vision-simulator"
  | "procedure-details";